pyproject.toml
[build-system]
requires  = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name  = "mytiktok"
version  = "0.0.5"
authors  = [
{name = "Dragoneyes", email =  "yolo89381@gmail.com"}]
description  = "A Simple Python Package for Scarping and Downloading Tiktok Videos"
readme  = "README.md"
classifiers  = [ 
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
    "attrs == 23.2.0",
  "backports.tarfile == 1.1.1",
  "blessings == 1.7",
  "blinker == 1.8.1",
  "Brotli == 1.1.0",
  "certifi == 2024.2.2",
  "cffi == 1.16.0",
  "charset-normalizer == 3.3.2",
  "cryptography == 42.0.5",
  "docutils == 0.21.2",
  "exceptiongroup == 1.2.1",
  "fake-useragent == 1.5.1",
  "h11 == 0.14.0",
  "h2 == 4.1.0",
  "hpack == 4.0.0",
  "hyperframe == 6.0.1",
  "idna == 3.7",
  "importlib_metadata == 7.1.0",
  "iniconfig == 2.0.0",
  "jaraco.classes == 3.4.0",
  "jaraco.context == 5.3.0",
  "jaraco.functools == 4.0.1",
  "jeepney == 0.8.0",
  "kaitaistruct == 0.10",
  "keyring == 25.2.0",
  "markdown-it-py == 3.0.0",
  "mdurl == 0.1.2",
  "more-itertools == 10.2.0",
  "nh3 == 0.2.17",
  "outcome == 1.3.0.post0",
  "packaging == 24.0",
  "pathvalidate == 3.2.0",
  "pkginfo == 1.10.0",
  "pluggy == 1.5.0",
  "pyasn1 == 0.6.0",
  "pycparser == 2.22",
  "Pygments == 2.17.2",
  "pyOpenSSL == 24.1.0",
  "pyparsing == 3.1.2",
  "PySocks == 1.7.1",
  "pytest == 8.2.1",
  "python-dotenv == 1.0.1",
  "readme_renderer == 43.0",
  "requests == 2.31.0",
  "requests-toolbelt == 1.0.0",
  "rfc3986 == 2.0.0",
  "rich == 13.7.1",
  "SecretStorage == 3.3.3",
  "selenium == 4.20.0",
  "six == 1.16.0",
  "sniffio == 1.3.1",
  "sortedcontainers == 2.4.0",
  "tenacity == 8.2.3",
  "tomli == 2.0.1",
  "tqdm == 4.66.4",
  "trio == 0.25.0",
  "trio-websocket == 0.11.1",
  "typing_extensions == 4.11.0",
  "undetected-chromedriver == 3.5.5",
  "urllib3 == 2.2.1",
  "websockets == 12.0",
  "wsproto == 1.2.0",
  "zipp == 3.18.1",
  "zstandard == 0.22.0"
]

[tool.setuptools]
# ...
# By default, include-package-data is true in pyproject.toml, so you do
# NOT have to specify this line.

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Homepage = "https://github.com/geovanigaldemsugar/Mytiktok"

